Replace hardcoded colors with new faces
This improves the presentation for all themes and gives power to the
user to change the applicable colors (or face attributes in general).

We inherit from `success' and `error' because they are always
available and should be supported by all themes.  The semantics
of those words do not apply in our case.